feat: add loading and empty state components for user management
- Introduced LoadingState component to display loading animation while fetching users.
- Added EmptyState component to show a message when no contacts are found.
- Updated utils to handle role icons with string type.
- Created userService for managing user CRUD operations.
- Refactored Sidebar to include a reusable ProfileDropdown component.
