+++ app/agenda/components/tabla.tsx
@@ -3,7 +3,7 @@ import {
   Phone, Mail, MoreHorizontal, Trash2, Edit, Eye
 } from "lucide-react"
 import { Badge } from "@/components/ui/badge"
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
+import { Avatar, AvatarImage } from "@/components/ui/avatar"
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table"
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@/components/ui/dropdown-menu"
 import { getRoleIcon, getStatusColor } from "./utils"
